AOC.Challenge.Day05

Description

Day 5. See AOC.Solver for the types used in this module!

After completing the challenge, it is recommended to:

Replace AOC.Prelude imports to specific modules (with explicit imports) for readability.

Remove the -Wno-unused-imports and -Wno-unused-top-binds pragmas.

Replace the partial type signatures underscores in the solution types _ :~> _ with the actual types of inputs and outputs of the solution. You can delete the type signatures completely and GHC will recommend what should go in place of the underscores.

Documentation

day05a :: Memory :~> Int

day05b :: Memory :~> Int

fillModes :: (Traversable t, Applicative t) => Int -> t Mode

Magically fills a fixed-shape Applicative with each mode from a mode op int.