AOC.Common.Subset

Documentation

findSubset

Arguments

:: (Monad m, Ord a)

=> (Set a -> m Ordering)
tester

-> Bool
whether or not to include empty set and full set

-> Set a
full set of items

-> m (Maybe (Set a))
subset that matches tester

testFinder

Arguments

:: Bool
whether or not to include empty set and full set

-> Int

-> Map (Set Int) Int

Get the number of guesses needed for each possible subset, for n items.

allBranches :: Ord a => Bool -> Set a -> [(Set a, [(Set a, Ordering)])]

buildDTree :: Ord a => Bool -> Set a -> Maybe (DTree a)

renderBranches :: (Ord a, Show a) => Bool -> Set a -> Maybe Text

renderBranchesChar :: Bool -> Set Char -> Maybe Text